refactor: make the ui module the core part
